code-analyzer/requirements.txt
grpcio==1.59.0
grpcio-tools==1.59.0
protobuf==4.24.4
tree-sitter==0.20.4
tree-sitter-python==0.23.6
tree-sitter-javascript==0.23.1
tree-sitter-java==0.23.5
tree-sitter-go==0.23.4
python-dotenv==1.0.0
